requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Model Integrations
openai==1.10.0
anthropic==0.18.0
# Optional: Uncomment for Hugging Face models
# transformers==4.37.0
# torch>=2.0.0
# huggingface-hub==0.20.3

# Code Execution & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-timeout==2.2.0
docker==7.0.0
coverage==7.4.0

# Code Quality Analysis
pylint==3.0.3
ruff==0.1.14
black==24.1.1
mypy==1.8.0
radon==6.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0

# Visualization & Dashboard
streamlit==1.30.0
plotly==5.18.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2

# Utilities
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
tqdm==4.66.1
tenacity==8.2.3

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1
asyncio==3.4.3

# Development
ipython==8.20.0
jupyter==1.0.0
pre-commit==3.6.0
